initial commit
The basic interpreter works.
The command line tool reads from stdin and outputs to stdout
Option for pretty ask for input, which can only be enabled/disabled by changing the source code for now.
Tests covering the internales
Test for a cat program that exits on NULL
Example of said cat program